Derive AudioScheduledSourceNode
