+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["servo-media", "examples"]
+members = ["servo-media", "servo-media-derive", "examples"]
 license = "MPL-2.0"
