+++ servo-media-derive/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "servo_media_derive"
+version = "0.1.0"
+authors = ["Fernando Jiménez Moreno <ferjmoreno@gmail.com>"]
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "0.11.11"
+quote = "0.3.15"
